+++ make/canvas.c2w.list-ucsver-detection-codes.sh
@@ -43,5 +43,36 @@ function list-range-code-for-version-detection {
     done
   done
 
+  printf '               | %-*s|musl\n' "$((nversion*3))" '-----Unicode EAW+GeneralCategory'
+  local -a keys=(
+    U+9FBC  U+9FC4  U+31B8  U+D7B0
+    U+3099  U+9FCD  U+1F93B U+312E
+    U+312F  U+16FE2 U+32FF  U+31BB
+    U+9FFD  U+1B132)
+  local code index=0 ret
+  for code in "${keys[@]}"; do
+    ((code=16#${code#U+}))
+    ble/unicode/EmojiStatus "$code"
+    if ((ret)); then
+      printf 'U+%04X: emoji cannot be used to detect the Unicode version\n' "$code" >&2
+      continue
+    fi
+
+    local c2w=${_ble_unicode_c2w[code]}
+    if [[ ! $c2w ]]; then
+      local c=$code
+      until [[ $c2w || c -eq 0 ]]; do c2w=${_ble_unicode_c2w[--c]}; done
+      if [[ $c2w ]]; then
+        printf 'U+%04X: warning: not c2w boundary. borrow the data of boundary U+%04X\n' "$code" "$c" >&2
+      else
+        printf 'U+%04X: this is not c2w boundary\n' "$code" >&2
+        continue
+      fi
+    fi
+
+    local width_vec=$(printf ' %2d' "${_ble_unicode_c2w_UnicodeVersionMapping[@]:c2w*nversion:nversion}")
+    ble/util/c2w:musl "$code"; local c2w_musl=$ret
+    printf '%-6s U+%05X |%s |%2d\n' "ws[$((index++))]" "$code" "$width_vec" "$c2w_musl"
+  done
 }
 list-range-code-for-version-detection